Indev focus fix

+++ src/lv_core/lv_style.c
@@ -31,7 +31,7 @@
  *  STATIC PROTOTYPES
  **********************/
 #if LV_USE_ANIMATION
-static void style_animator(lv_style_anim_dsc_t * dsc, int32_t val);
+static void style_animator(lv_style_anim_dsc_t * dsc, int16_t val);
 static void style_animation_common_end_cb(lv_anim_t * a);
 #endif
 
@@ -316,9 +316,9 @@ void lv_style_anim_set_styles(lv_anim_t * a, lv_style_t * to_anim, const lv_styl
 /**
  * Used by the style animations to set the values of a style according to start and end style.
  * @param dsc the 'animated variable' set by lv_style_anim_create()
- * @param val the current state of the animation between 0 and LV_STYLE_ANIM_RES
+ * @param val the current state of the animation between 0 and LV_ANIM_RESOLUTION
  */
-static void style_animator(lv_style_anim_dsc_t * dsc, int32_t val)
+static void style_animator(lv_style_anim_dsc_t * dsc, int16_t val)
 {
     const lv_style_t * start = &dsc->style_start;
     const lv_style_t * end   = &dsc->style_end;
